// generated/src/app/api-models/auth-procedure-range-value.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AuthProcedureRangeValue {
  authProcRangeId: string;
  fromRange: string;
  thruRange: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}